% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance.R
\name{fetch_replicate_weights}
\alias{fetch_replicate_weights}
\title{Fetch the long-format replicate weights}
\usage{
fetch_replicate_weights(conn)
}
\arguments{
\item{conn}{A \code{DBIConnection} from \code{\link[=lsa_connect]{lsa_connect()}}.}
}
\value{
A data.frame with \code{student_id}, \code{replicate_id}, \code{weight}.
}
\description{
Fetch the long-format replicate weights
}
