% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance.R
\name{fetch_replicates}
\alias{fetch_replicates}
\title{Fetch the replicate definitions}
\usage{
fetch_replicates(conn)
}
\arguments{
\item{conn}{A \code{DBIConnection} from \code{\link[=lsa_connect]{lsa_connect()}}.}
}
\value{
A data.frame with \code{replicate_id}, \code{canton}, \code{jk_factor}.
}
\description{
Fetch the replicate definitions
}
